requirements.txt
jieba==0.42.1
Pillow==8.2.0
